backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
PyMuPDF==1.23.21
python-docx==1.1.0
openpyxl==3.1.2
spacy==3.7.2
chardet==5.2.0
structlog==24.1.0
hypothesis==6.98.3
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0
